removed unnecessary files
