--- HNG-00597.js
@@ -1,15 +1,0 @@
-const Name = "Christian Dirisu";
-
-const ID = "HNG-00597";
-
-const Language = "Javascript";
-
-const Email = "christian.dirisu@gmail.com";
-
-const task = () => {
-
-    return `Hello world, this is ${Name} with HNGi7 ID ${ID} and email ${Email} using ${Language} for stage 2 task`
-
-};
-
-console.log(task());
