--- HNG-04274.js
@@ -1,5 +1,0 @@
-const myFullName = `Ogbu Ezekiel Ike`;
-const myHNGDetails =  `HNGi7 ID HNG-04274`;
-const programmingLanguageUsed = `Javascript`;
-const myEmail = `Zicozydasilva70@gmail.com`
-console.log(`Hello World, this is ${myFullName} with ${myHNGDetails} using ${programmingLanguageUsed} for stage 2 task.. ${myEmail}`);
