--- scripts/HNG-00442.py
@@ -1,9 +1,0 @@
-Name = "Jayasree Kurmannagari"
-Id = "HNG-00442"
-emailId = "kjsree1998@gmail.com"
-language = "Python"
- 
-def hello():
-  return "Hello World, this is " + Name + " with HNGi7 ID " + Id + " using " + language + " for stage 2 task. " + emailId
-
-print(hello(), flush = True)
