--- scripts/HNG-00869.py
@@ -1,8 +1,0 @@
-#HNG Task-2 for team falcon
-
-
-Hng_Details = input(
-    "Hello world this is Adeyemi Oluwatobi with HNGi7 ID of HNG-00869, using python language for stage 2 task. tobinuel155@gmail.com")
-
-print(Hng_Details)
-
