+++ scripts/HNG-06429.py
