--- scripts/mytaskC#/Program.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace mytaskC_
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            Console.WriteLine("Hello World, this is Osakinle Dare Thomas with HNGi7 ID 00937 using C# for stage 2 task. Thomasdreh@gmail.com");
-        }
-    }
-}
