--- scripts/mytaskC#/bin/Debug/netcoreapp3.1/mytaskC#.deps.json
@@ -1,23 +1,0 @@
-{
-  "runtimeTarget": {
-    "name": ".NETCoreApp,Version=v3.1",
-    "signature": ""
-  },
-  "compilationOptions": {},
-  "targets": {
-    ".NETCoreApp,Version=v3.1": {
-      "mytaskC#/1.0.0": {
-        "runtime": {
-          "mytaskC#.dll": {}
-        }
-      }
-    }
-  },
-  "libraries": {
-    "mytaskC#/1.0.0": {
-      "type": "project",
-      "serviceable": false,
-      "sha512": ""
-    }
-  }
-}
