--- scripts/mytaskC#/bin/Debug/netcoreapp3.1/mytaskC#.runtimeconfig.dev.json
@@ -1,8 +1,0 @@
-{
-  "runtimeOptions": {
-    "additionalProbingPaths": [
-      "C:\\Users\\USER\\.dotnet\\store\\|arch|\\|tfm|",
-      "C:\\Users\\USER\\.nuget\\packages"
-    ]
-  }
-}
