--- scripts/mytaskC#/bin/Debug/netcoreapp3.1/mytaskC#.runtimeconfig.json
@@ -1,9 +1,0 @@
-{
-  "runtimeOptions": {
-    "tfm": "netcoreapp3.1",
-    "framework": {
-      "name": "Microsoft.NETCore.App",
-      "version": "3.1.0"
-    }
-  }
-}
